refactor: planned event page
